Move CI/CD from TravisCI to Github Actions
Since the free plan on TravisCI is ending June 29, 2025, we should move over our CI/CD to Github Actions which is free for public repositories.

Add `Build` and `Release` workflows.
- The `Build` workflow runs on any push to the `main` branch, or on any pull-request change, and will install dependencies, run the tests, ensure 100% coverage for the project and changes, and run the linter.
- The `Release` workflow runs any time we publish a release. It will build the release distributions (sdist and wheel) as part of the first job, and then publish those distributions to PyPI as part of the second job.

Also bumped up the Python version supported from 3.7 - 3.9 to 3.10 - 3.12, since 3.7 and 3.8 is EOL, and 3.9 will go EOL in October 2025.

Fixes #40